requirements.txt
colivara_py==0.1.2
numpy==2.1.3
pandas==2.2.3
pytest==8.3.3
python-dotenv==1.0.1
tqdm==4.66.5
pytest-mock